Gravatar added to users with email addresses confirmed.
Hummm. I had it in mind that we'd just display the user's
email but that's just asking for spam isn't it? Perhaps
we'll have to make avatars clickable to send PMs or
something? Just redirect some email, or give each
user a nice PM page or something?

Leaving it merely not displayed for now.