Now works cross-domain.
Woah, that was a struggle there for a while, I had no idea
about the HTTP "OPTIONS" request or the accept-* headers
that are needed. Scary, though it was gonna just ban
me from doing this for a second.

But it's in now, and working, and sending the cookies,
and everything's lovely.