Added system for viewing user's past messages.
Fairly complex, you pass it a message and a nickname
and it'll show just users in that session with that
nick, unless that session has an email address, in
which case *all* sessions with that email.

If you supply a nick of -1 it'll then branch out
to all nicks in that cookie (or cookies with that
email).