Override the CSRF protection with something different.
We want to check CSRF attacks, certainly, but we want to
allow the user to send in requests from different
browser windows all running simultaneously. The old
CSRF protection doesn't handle this well. This one
should be better, using the date modulo 10 minutes
as a key (combined with session cookies etc).