Fixed up all the default test.
We now have a test suite of sorts, and it passes all tests!
WoO!

These tests don't mean a lot other than that the
pages return something vaguely right though, and
there's no model tests at all still!

But nevertheless, it's a bunch of phpunit tests
in Zend and it's the first time I've done anything
like that.